--- web/apps/photos/src/components/FixCreationTime/running.tsx
@@ -1,32 +1,0 @@
-import { LinearProgress } from "@mui/material";
-import { ComfySpan } from "components/ExportInProgress";
-import { t } from "i18next";
-
-export default function FixCreationTimeRunning({ progressTracker }) {
-    const progress = Math.round(
-        (progressTracker.current * 100) / progressTracker.total,
-    );
-    return (
-        <>
-            <div style={{ marginBottom: "10px" }}>
-                <ComfySpan>
-                    {" "}
-                    {progressTracker.current} / {progressTracker.total}{" "}
-                </ComfySpan>{" "}
-                <span style={{ marginLeft: "10px" }}>
-                    {" "}
-                    {t("CREATION_TIME_UPDATED")}
-                </span>
-            </div>
-            <div
-                style={{
-                    width: "100%",
-                    marginTop: "10px",
-                    marginBottom: "20px",
-                }}
-            >
-                <LinearProgress variant="determinate" value={progress} />
-            </div>
-        </>
-    );
-}
